add keypair generation to device registration
